// src/model/courseContentResponseDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * UBUMonitor-Analytics
 * API documentation for UBUMonitor Analytics
 * OpenAPI spec version: 1.0.0
 */
import type { CourseContentSectionDto } from './courseContentSectionDto';

export interface CourseContentResponseDto {
  sections: CourseContentSectionDto[];
}