// src/model/courseContentSectionDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * UBUMonitor-Analytics
 * API documentation for UBUMonitor Analytics
 * OpenAPI spec version: 1.0.0
 */
import type { CourseContentModuleDto } from './courseContentModuleDto';

export interface CourseContentSectionDto {
  /** Section ID */
  id: number;
  /** Section name */
  name: string;
  /** Is the section visible */
  visible: boolean;
  /** Section description */
  summary: string;
  /** Section number inside the course */
  position: number;
  /** List of modules in the section */
  modules: CourseContentModuleDto[];
}