// src/model/courseEnrollmentsResponseDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * UBUMonitor-Analytics
 * API documentation for UBUMonitor Analytics
 * OpenAPI spec version: 1.0.0
 */
import type { UserEnrollmentInfoDto } from './userEnrollmentInfoDto';

export interface CourseEnrollmentsResponseDto {
  /** List of users and their course enrollment information */
  users: UserEnrollmentInfoDto[];
}