// src/model/userEnrollmentInfoDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * UBUMonitor-Analytics
 * API documentation for UBUMonitor Analytics
 * OpenAPI spec version: 1.0.0
 */
import type { UserEnrollmentInfoCoursesInnerDto } from './userEnrollmentInfoCoursesInnerDto';
import type { UserEnrollmentInfoGroupsInnerDto } from './userEnrollmentInfoGroupsInnerDto';
import type { UserEnrollmentInfoRolesInnerDto } from './userEnrollmentInfoRolesInnerDto';

export interface UserEnrollmentInfoDto {
  /** The unique identifier of the user. */
  id: number;
  /** The full name of the user. */
  fullName: string;
  /** The email address of the user. */
  email: string;
  /** The timestamp of the user's first access to the site. */
  firstAccess: string;
  /** The timestamp of the user's last access to the site. */
  lastAccess: string;
  /** The timestamp of the user's last access to the course. */
  lastCourseAccess: string;
  /** The URL of the user's profile image. */
  imageUrl: string;
  /** List of roles the user has in the course. */
  roles: UserEnrollmentInfoRolesInnerDto[];
  /** List of groups the user is enrolled in for the course. */
  groups: UserEnrollmentInfoGroupsInnerDto[];
  /** List of courses the user is enrolled in. */
  courses: UserEnrollmentInfoCoursesInnerDto[];
}